refactor(css modules): rename files to match the component file name
