+++ src/modules/CardMenu/CardMenu.js
@@ -4,7 +4,7 @@ import React from 'react'
 import { List } from '../icons/List'
 import { MenuCard, ActionWrapper } from '../MenuCard/MenuCard'
 import { Plus } from '../icons/Plus'
-import styles from './styles.module.css'
+import styles from './CardMenu.module.css'
 
 export const CardMenu = ({ sections }) => {
     const filteredSections = sections.filter(
